feat: Separate useResize hook
